+++ spec/lib/xe_client/models/quote_spec.rb
@@ -7,6 +7,7 @@ module XEClient
       subject { described_class }
       it { is_expected.to have_attribute(:quotecurrency, String) }
       it { is_expected.to have_attribute(:mid, Float) }
+      it { is_expected.to have_attribute(:timestamp, DateTime) }
     end
 
   end